Full rework

+++ include/stdlib/mem.h
@@ -1,12 +1,13 @@
 #ifndef STDLIB
 #define STDLIB
 
+#include <ntypes.h>
 #include <stddef.h>
 
-void *malloc(unsigned long long size);
+void *malloc(u64 size);
 
-void *memcpy(void *dst, void* src, unsigned long long n);
-void *memset(void *dst, int c, unsigned long long n);
-void *bzero(void *dst, unsigned long long n);
+void *memcpy(void *dst, void *src, u64 n);
+void *memset(void *dst, u8 c, u64 n);
+void *bzero(void *dst, u64 n);
 
 #endif
