+++ include/syslib/proc.h
@@ -1,12 +1,13 @@
 #ifndef PROC_H
 #define PROC_H
 
+#include <ntypes.h>
 
 #define NCPU 64
 
 struct cpu {
-    int cli_depth;
-    int intr_enabled;
+    u8 cli_depth;
+    u8 intr_enabled;
 };
 
 struct cpu *getcpu(void);
